IDL parser: implement Stringifier production
Emit a node with class 'Stringifier' for all stringifier variants.
If the stringifier is an attribute, this node has a child node that
represents the attribute. If the stringifier is an operation (named
or unnamed) the node has a child node that represents the operation.

BUG=306606

Review URL: https://codereview.chromium.org/329163002

git-svn-id: http://src.chromium.org/chrome/trunk/src/tools/idl_parser@279097 4ff67af0-8c30-449e-8e8b-ad334ec8d88c